BackEnd 3.0/requirements.txt
cffi==1.15.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
cryptography==38.0.3
Deprecated==1.2.13
dnspython==2.2.1
Flask==2.2.2
Flask-Limiter==2.7.0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
limits==2.7.1
MarkupSafe==2.1.1
packaging==21.3
pycparser==2.21
pycryptodomex==3.15.0
Pygments==2.13.0
pymongo==4.3.2
pyparsing==3.0.9
rich==12.6.0
typing_extensions==4.4.0
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.10.0
